requirements.txt
async-timeout==4.0.3
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
spotipy==2.23.0
urllib3==2.0.7
